- update BulkUpdate fn
